+++ example/demo-ui-tools.js
@@ -2,6 +2,14 @@
 // ==============================================================
 // UI manipulation for demo purposes, for ElvenJS see below
 // ==============================================================
+export const uiPending = (isLoading, button) => {
+  const overlay = document.querySelector('.overlay');
+  if (isLoading) {
+    overlay.classList.add('visible');
+  } else {
+    overlay.classList.remove('visible');
+  }
+};
 
 export const uiLoggedInState = (loggedIn) => {
   const loginExtensionButton = document.getElementById(
@@ -12,6 +20,7 @@ export const uiLoggedInState = (loggedIn) => {
   const txButton = document.getElementById('button-tx');
   const txEsdtButton = document.getElementById('button-tx-esdt');
   const mintButton = document.getElementById('button-mint');
+  const queryButton = document.getElementById('button-query');
   if (loggedIn) {
     loginExtensionButton.style.display = 'none';
     loginMaiarButton.style.display = 'none';
@@ -19,6 +28,7 @@ export const uiLoggedInState = (loggedIn) => {
     txButton.style.display = 'block';
     txEsdtButton.style.display = 'block';
     mintButton.style.display = 'block';
+    queryButton.style.display = 'block';
   } else {
     loginExtensionButton.style.display = 'block';
     loginMaiarButton.style.display = 'block';
@@ -26,73 +36,43 @@ export const uiLoggedInState = (loggedIn) => {
     txButton.style.display = 'none';
     txEsdtButton.style.display = 'none';
     mintButton.style.display = 'none';
+    queryButton.style.display = 'none';
   }
-};
-
-export const uiSpinnerState = (isLoading, button) => {
-  const buttonLoginExtension = document.getElementById(
-    'button-login-extension'
-  );
-  const buttonLoginMobile = document.getElementById('button-login-mobile');
-  const buttonEgld = document.getElementById('button-tx');
-  const buttonEsdt = document.getElementById('button-tx-esdt');
-  const buttonMint = document.getElementById('button-mint');
-  const spinnerText = 'Transaction pending...';
-  if (isLoading) {
-    if (button === 'loginExtension') {
-      buttonLoginExtension.innerText = 'Logging in...';
-      buttonLoginExtension.setAttribute('disabled', true);
-    }
-    if (button === 'loginMobile') {
-      buttonLoginMobile.innerText = 'Logging in...';
-      buttonLoginMobile.setAttribute('disabled', true);
-    }
-    if (button === 'egld') {
-      buttonEgld.innerText = spinnerText;
-      buttonEgld.setAttribute('disabled', true);
-    }
-    if (button === 'esdt') {
-      buttonEsdt.innerText = spinnerText;
-      buttonEsdt.setAttribute('disabled', true);
-    }
-    if (button === 'mint') {
-      buttonMint.innerText = spinnerText;
-      buttonMint.setAttribute('disabled', true);
-    }
-  } else {
-    if (button === 'loginExtension') {
-      buttonLoginExtension.innerText = 'Login with Extension';
-      buttonLoginExtension.removeAttribute('disabled');
-    }
-    if (button === 'loginMobile') {
-      buttonLoginMobile.innerText = 'Login with Maiar mobile';
-      buttonLoginMobile.removeAttribute('disabled');
-    }
-    if (button === 'egld') {
-      buttonEgld.innerText = 'EGLD transaction';
-      buttonEgld.removeAttribute('disabled');
-    }
-    if (button === 'esdt') {
-      buttonEsdt.innerText = 'ESDT transaction*';
-      buttonEsdt.removeAttribute('disabled');
-    }
-    if (button === 'mint') {
-      buttonMint.innerText = 'Mint NFT';
-      buttonMint.removeAttribute('disabled');
-    }
-  }
+  uiPending(false);
 };
 
 export const updateTxHashContainer = (txHash) => {
-  const txHashContainer = document.getElementById('tx-hash');
+  const txHashContainer = document.getElementById('tx-hash-or-query-result');
   if (txHash) {
+    txHashContainer?.replaceChildren();
     const url = `https://devnet-explorer.elrond.com/transactions/${txHash}`;
     const link = document.createElement('a');
     link.setAttribute('href', url);
     link.classList.add('transaction-link');
-    link.innerText = url;
-    txHashContainer.appendChild(link);
+    link.innerText = `➡️ ${url}`;
+    txHashContainer?.appendChild(link);
   } else {
-    txHashContainer?.querySelector('a')?.remove();
+    txHashContainer?.replaceChildren();
   }
+};
+
+export const updateQueryResultContainer = (result) => {
+  const queryContainer = document.getElementById('tx-hash-or-query-result');
+  if (result) {
+    queryContainer?.replaceChildren();
+    const divElem = document.createElement('div');
+    divElem.innerText = result;
+    queryContainer?.appendChild(divElem);
+  } else {
+    queryContainer?.replaceChildren();
+  }
+};
+
+// For simplicity, you should probably use Buffer in browser for that
+export const base64ToDecimalHex = (str) => {
+  const raw = window.atob(str);
+  const result = [...raw].map((c) =>
+    c.charCodeAt(0).toString(16).padStart(2, 0)
+  ).join``;
+  return result.toUpperCase();
 };